+++ iatikit/data/dataset.py
@@ -207,13 +207,12 @@ class DatasetSet(GenericSet):
     """
 
     _key = 'name'
-    _filters = ['name', 'filetype', 'xpath']
+    _filters = ['name', 'filetype']
+    _multi_filters = ['xpath']
     _instance_class = Dataset
 
     def __init__(self, data_path, metadata_path, **kwargs):
-        super(DatasetSet, self).__init__()
-        self.wheres = kwargs
-
+        super(DatasetSet, self).__init__(**kwargs)
         self.data_path = data_path
         self.metadata_path = metadata_path
 
@@ -231,23 +230,27 @@ class DatasetSet(GenericSet):
                  for x in set(list(data_paths.keys()) +
                               list(metadata_paths.keys()))}
 
-        name = self.wheres.get('name')
-        if name is not None:
-            paths = [paths[name]] if name in paths else []
+        where_name = self.wheres.get('name')
+        if where_name is not None:
+            paths = [paths[where_name]] if where_name in paths else []
         else:
             paths = sorted(list(paths.values()))
 
         where_filetype = self.wheres.get('filetype')
-        where_xpath = self.wheres.get('xpath')
+        where_xpaths = self.wheres.get('xpath', [])
 
         for data_path, metadata_path in paths:
             dataset = Dataset(data_path, metadata_path)
             if where_filetype is not None and \
                     dataset.filetype != where_filetype:
                 continue
-            if where_xpath is not None:
+            if where_xpaths != []:
                 if not dataset.validate_xml():
                     continue
-                if dataset.etree.xpath(where_xpath) == []:
-                    continue
+                for where_xpath in where_xpaths:
+                    if dataset.etree.xpath(where_xpath) == []:
+                        break
+                else:
+                    yield dataset
+                continue
             yield dataset