+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.2.10] – 2020-07-31
 
 ### Fixed
 - Deal with missing XML files gracefully. ([#52](https://github.com/codeforIATI/iatikit/issues/52))